DPP: De-const have_voice option and add coroutines support. (#4824)
* fix(dpp): Remove repeated have_voice condition

* feat(dpp): Make have_voice option modifiable

De-const have_voice option and enables optional download of voice libraries

* feat(dpp): Add coroutine support

Adds a configurable option to enable coroutine support

* fix(dpp)!: Change have_voice to voice inside package

As mentioned in the same pull request, prefixes like have_ are discouraged in xmake packages.

BREAKING CHANGES: This WILL break xmake scripts that use the have_voice flag, IMO it's better to keep the previous have_voice flag as a deprecated alias of the new voice flag, and in the next version of D++ completely remove it from the package.

* fix(dpp): Add C++ minimum version for test

If coroutines are enabled, C++20 is required or the compilation will fail even though the test only requires C++17 to work, as dpp automatically includes coroutines if the macro is defined without checking C++ version

* fix(dpp): Set C++ language for library dinamically

The current configuration wasn't proper and when coroutines were enabled, the C++ version would change to C++17 sporadically, ignoring the later C++20 specification that was done later. Now the C++ version will adjust correctly to the coro flag.

* style(dpp): Remove requested empty lines

* feat(dpp): Re-add have_voice flag with deprecated warning

This will bring up again have_voice to avoid codebase breakage, but with many deprecation warnings around to prevent new users to define it.