Fix Intel compiler flags for OpenMP (#7540)
* Fix Intel compiler flags for OpenMP

The Intel C++ Compiler Classic (`icc`) is deprecated and was discontinued 2024.
As an alternative, it is recommended that developers use the Intel OneAPI C++ compiler (`icx`).
The same is true for the Intel Fortran compiler: `ifort` is deprecated in favour of `ifx`.
See the [icc Release notes](https://www.intel.com/content/www/us/en/developer/articles/release-notes/oneapi-c-compiler-release-notes.html) for more details.

Following the name change, xmake failed to provide the correct compiler flags for OpenMP, a parallel computing library for multicore processors.
See also xmake-io/xmake#6596 for more information.

This pull request involves making minor changes to the xmake configuration for OpenMP, essentially renaming 'icc' to 'icx' and 'ifort' to 'ifx', to ensure that the Intel compiler is correctly detected and that OpenMP receives the correct compiler flags.

* keep old and new

---------

Co-authored-by: Mavoort <mavoort@posteo.org>
Co-authored-by: Saikari <lin@sz.cn.eu.org>